+++ edgeworkers/New-EdgeWorkerAuthToken.ps1
@@ -1,44 +1,38 @@
 function New-EdgeWorkerAuthToken
 {
-    [CmdletBinding(DefaultParameterSetName = 'url')]
+    [CmdletBinding(DefaultParameterSetName = 'hostname')]
     Param(
-        [Parameter(Mandatory=$true)]  [string] $Secret,
-        [Parameter(Mandatory=$true,ParameterSetName='acl')] [string] $ACLPath,
-        [Parameter(Mandatory=$true,ParameterSetName='url')] [string] $URLPath,
-        [Parameter(Mandatory=$false)] [int] $Expiry = 15
+        [Parameter(Mandatory=$true,ParameterSetName='hostname')]   [string] $Hostname,
+        [Parameter(Mandatory=$true,ParameterSetName='hostnames')]  [string] $Hostnames,
+        [Parameter(Mandatory=$true,ParameterSetName='propertyid')] [string] $PropertyID,
+        [Parameter(Mandatory=$false)] [int] $Expiry = 480,
+        [Parameter(Mandatory=$false)] [string] $EdgeRCFile,
+        [Parameter(Mandatory=$false)] [string] $Section,
+        [Parameter(Mandatory=$false)] [string] $AccountSwitchKey
     )
 
-    $Delimiter = '~'
-    $Token = ''
-    $Start = [long] (Get-Date -Date ((Get-Date).ToUniversalTime()) -UFormat %s)
-    $End = $Start + ($Expiry * 60)
-
-    $Token += "st=$Start"
-    $Token += $Delimiter
-    $Token += "exp=$End"
-
-    $HMACToken = $Token
-    $HMACToken += $Delimiter
-    if($ACLPath){
-        $Token += $Delimiter
-        $Token += "acl=$ACLPath"
-        $HMACToken = $Token
-    }
-    else{
-        $HMACToken = $Token
-        $HMACToken += $Delimiter
-        $EncodedURL = [System.Web.HttpUtility]::UrlEncode($URLPath)
-        $HMACToken += "url=$EncodedURL"
+    $Path = "/edgeworkers/v1/secure-token"
+    $BodyObj = [PSCustomObject] @{
+        expiry = $Expiry
     }
 
-    # Generate HMAC
-    $HMACSHA = New-Object System.Security.Cryptography.HMACSHA256
-    $HMACSHA.key = [byte[]] -split ($Secret -replace '..', '0x$& ') # Secret is presented as string, but we need to treat each pair of characters as Hex before getting their byte value
-    $Hash = $HMACSHA.ComputeHash([Text.Encoding]::UTF8.GetBytes($HMACToken))
-    $Signature = ($Hash | foreach ToString x2) -join ''
+    if($PSCmdlet.ParameterSetName -eq 'hostname'){
+        $BodyObj | Add-Member -MemberType NoteProperty -Name 'hostname' -Value $Hostname
+    }
+    elseif($PSCmdlet.ParameterSetName -eq 'hostnames'){
+        $BodyObj | Add-Member -MemberType NoteProperty -Name 'hostnames' -Value ($Hostnames -split ',')
+    }
+    elseif($PSCmdlet.ParameterSetName -eq 'propertyid'){
+        $BodyObj | Add-Member -MemberType NoteProperty -Name 'propertyId' -Value $PropertyID
+    }
 
-    $Token += $Delimiter
-    $Token += "hmac=$Signature"
+    $Body = ConvertTo-Json $BodyObj
 
-    return $Token
+    try {
+        $Result = Invoke-AkamaiRestMethod -Method POST -Path $Path -Body $Body -EdgeRCFile $EdgeRCFile -Section $Section -AccountSwitchKey $AccountSwitchKey
+        return $Result.akamaiEwTrace
+    }
+    catch {
+        throw $_
+    }
 }